イベントモデルのenumをコンサーンに移動
